requirements.txt
# Core dependencies
psycopg2-binary>=2.9.7
streamlit>=1.28.0

# AI/ML dependencies
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Optional: For Llama models (uncomment if using)
# torch-audio>=2.0.0
# torchaudio>=2.0.0

# Google Gemini API
google-generativeai>=0.3.0

# Optional: OpenAI API (uncomment if using)
# openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
feedparser>=6.0.10
requests>=2.31.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# RAG and Vector Database
pathway>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Web Framework
flask>=2.3.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
